Move CSS imports to dedicated styles directory
